services/request_site_interface/requirements.txt
fastapi==0.124.0
sqlalchemy==2.0.44
uvicorn==0.38.0
gunicorn==23.0.0
asyncpg==0.31.0
kafka-python==2.3.0
aiohttp==3.13.2